refactor: add validate key logic & Improved folder structure
