+++ packages/utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@log/utils",
+  "name": "@clarity/utils",
   "version": "0.0.1",
   "main": "lib/index.js",
   "scripts": {
@@ -9,7 +9,9 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "dependencies": {},
+  "dependencies": {
+    "@clarity/types": "workspace:*"
+  },
   "devDependencies": {
     "@types/node": "18.11.18",
     "typescript": "5.3.3"
