+++ packages/utils/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./logger";
+export * from "./consoleLogger";
+export * from "./isValidSecretKey";
