+++ packages/utils/src/isValidSecretKey.ts
@@ -1,0 +1,4 @@
+export function isValidSecretKey(secretKey: string): Boolean {
+  // TODO: api check
+  return Boolean(secretKey);
+}
