--- packages/utils/src/logger/index.ts
@@ -1,2 +1,0 @@
-export { Logger } from "./logger";
-export type { LoggerService } from "./interface";
